refactor(events) : place event types in event class for consistency
