+++ app/src/main/java/com/android/unio/model/event/Event.kt
@@ -1,8 +1,17 @@
 package com.android.unio.model.event
 
+import androidx.compose.ui.graphics.Color
 import com.android.unio.model.association.Association
 import com.android.unio.model.firestore.ReferenceList
 import com.android.unio.model.map.Location
+import com.android.unio.ui.theme.eventTypeAperitif
+import com.android.unio.ui.theme.eventTypeFestival
+import com.android.unio.ui.theme.eventTypeJam
+import com.android.unio.ui.theme.eventTypeNetworking
+import com.android.unio.ui.theme.eventTypeNightParty
+import com.android.unio.ui.theme.eventTypeOther
+import com.android.unio.ui.theme.eventTypeSport
+import com.android.unio.ui.theme.eventTypeTrip
 import com.google.firebase.Timestamp
 import java.util.Date
 
@@ -34,6 +43,17 @@ data class Event(
     val date: Timestamp = Timestamp(Date()),
     val location: Location = Location(),
     val types: List<EventType> = mutableListOf(EventType.OTHER)
-) {
-  companion object
+){
+    companion object
+}
+
+enum class EventType(val color: Color, val text: String) {
+    FESTIVAL(eventTypeFestival, "festival"),
+    APERITIF(eventTypeAperitif, "aperitif"),
+    NIGHT_PARTY(eventTypeNightParty, "night party"),
+    JAM(eventTypeJam, "jam"),
+    NETWORKING(eventTypeNetworking, "networking"),
+    SPORT(eventTypeSport, "sport"),
+    TRIP(eventTypeTrip, "trip"),
+    OTHER(eventTypeOther, "other"); // Default color
 }
